+++ test/api_tests/matrixMul/memory/MatrixAllocator.h
@@ -13,7 +13,7 @@
 
 class MatrixAllocator : public htgs::IMemoryAllocator<double *> {
  public:
-  MatrixAllocator(int width, int height) : IMemoryAllocator((size_t) width * height) { }
+  MatrixAllocator(size_t width, size_t height) : IMemoryAllocator((size_t) width * height) { }
 
   double *memAlloc(size_t size) {
     double *mem = new double[size];
